Move serialization code into a separate module
