+++ c/src/lib.rs
@@ -4,3 +4,4 @@ pub mod util;
 pub mod atom;
 pub mod space;
 pub mod metta;
+pub mod serial;
